Changed paths and moved file
Changed paths to match with the task and moved model.py to properly build .whl with poetry
