+++ pyproject.toml
@@ -6,6 +6,7 @@ authors = [
     {name = "Anisimov Aleksey",email = "fallengalaxy5000@gmail.com"}
 ]
 readme = "README.md"
+packages = [{include = "machinelearning"}]
 requires-python = ">=3.11"
 dependencies = [
     "catboost (>=1.2.7,<2.0.0)",
